+++ CHANGELOG.md
@@ -9,6 +9,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.1.1] - 2021-01-20
+
+### Added
+
+- Add `DefaultGRPCPort` constant
+
+### Changed
+
+- `DNSTargetDiscoverer` now uses `DefaultGRPCPort` by default
+
 ## [0.1.0] - 2021-01-20
 
 - Initial release
@@ -16,6 +26,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 <!-- references -->
 [Unreleased]: https://github.com/dogmatiq/discoverkit
 [0.1.0]: https://github.com/dogmatiq/discoverkit/releases/tag/v0.1.0
+[0.1.1]: https://github.com/dogmatiq/discoverkit/releases/tag/v0.1.1
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD